Correcting test_files.py
